+++ Version1/src/main/java/org/example/DomainLayer/UserAggregate/User.java
@@ -16,11 +16,11 @@ public class User {
     private String passwordHash;
     private UserRole role;
     private UserStatus status;
-    private int age;
+    private float age;
     private HashMap<UUID, ICompanyMember> companyRoles;
     private final Map<UUID, Invitation> CompanyInvitations;
 
-    public User(UUID id, String username, String email, String passwordHash, int age) {
+    public User(UUID id, String username, String email, String passwordHash, float age) {
         this.id = id;
         this.username = username;
         this.email = email;
@@ -58,7 +58,7 @@ public class User {
                 String passwordHash,
                 UserRole role,
                 UserStatus status,
-                int age) {
+                float age) {
         this.id = id;
         this.username = username;
         this.email = email;
@@ -391,7 +391,7 @@ public class User {
         return status;
     }
 
-    public int getAge() {
+    public float getAge() {
         return age;
     }
